# openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "SalesAPI",
    "summary": "Sales REST API for managing customers, products, orders, and sales data.",
    "description": "A RESTful API for the Sales POC application. Provides CRUD operations for customers, products, sales orders, order items, and sales representatives, as well as read-only access to aggregated sales facts.",
    "version": "1.0.0"
  },
...
  "servers": [
    {
      "url": "https://salespoc-api.azurewebsites.net/"
    }
  ],
  "paths": {
...
    "/api/Products": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Get all products",
        "description": "Retrieves a list of all products.",
        "operationId": "getProducts",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Products"
        ],
        "summary": "Create a product",
        "description": "Creates a new product record.",
        "operationId": "createProduct",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Product"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Product"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Product"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          }
        }
      }
    },
    "/api/Products/{id}": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Get a product by ID",
        "description": "Retrieves a single product by its unique identifier.",
        "operationId": "getProductById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Products"
        ],
        "summary": "Update a product",
        "description": "Updates an existing product by its unique identifier.",
        "operationId": "updateProduct",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Product"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Product"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Product"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Products"
        ],
        "summary": "Delete a product",
        "description": "Deletes a product by its unique identifier.",
        "operationId": "deleteProduct",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/SalesFacts/product/{productName}": {
      "get": {
        "tags": [
          "SalesFacts"
        ],
        "summary": "Get sales facts by product name",
        "description": "Retrieves sales facts filtered by product name.",
        "operationId": "getSalesFactsByProduct",
        "parameters": [
          {
            "name": "productName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VwSalesFact"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VwSalesFact"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VwSalesFact"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/ProductDocuments/by-product/{productName}": {
      "get": {
        "tags": [
          "ProductDocuments"
        ],
        "summary": "Get documents by product",
        "description": "Lists all documents for a specific product (e.g. Chip-1). Each product typically has: Engineering DataSheet, Marketing Brief, Sales One Pager.",
        "operationId": "getDocumentsByProduct",
        "parameters": [
          {
            "name": "productName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The product name to filter by (e.g. Chip-1)"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductDocument"
                  }
                }
              }
            }
          },
          "404": {
            "description": "No documents found for the specified product"
          }
        }
      }
    },
    "/api/ProductDocuments/download": {
      "get": {
        "tags": [
          "ProductDocuments"
        ],
        "summary": "Download a document",
        "description": "Downloads a specific document by its full blob name from Azure Blob Storage.",
        "operationId": "downloadProductDocument",
        "parameters": [
          {
            "name": "blobName",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The full blob name of the document to download"
          }
        ],
        "responses": {
          "200": {
            "description": "The document file",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "description": "blobName query parameter is required"
          },
          "404": {
            "description": "Blob not found"
          }
        }
      }
    },
...
    "/api/ProductDescriptions/{productId}": {
      "get": {
        "tags": [
          "ProductDescriptions"
        ],
        "summary": "Get product description by ID",
        "description": "Retrieves a product description by its ID (e.g. Chip-100).",
        "operationId": "getProductDescriptionById",
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The product ID (e.g. Chip-100)"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDescription"
                }
              }
            }
          },
          "404": {
            "description": "No product description found for the specified ID"
          }
        }
      }
    },
    "/api/ProductDescriptions/search": {
      "get": {
        "tags": [
          "ProductDescriptions"
        ],
        "summary": "Search product descriptions",
        "description": "Searches product descriptions by partial ID match (case-insensitive).",
        "operationId": "searchProductDescriptions",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The search term to match against product IDs"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductDescription"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Query parameter q is required"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Customer": {
        "type": "object",
        "properties": {
          "customerId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "customerName": {
            "type": "string"
          },
          "customerType": {
            "type": [
              "null",
              "string"
            ]
          },
          "industry": {
            "type": [
              "null",
              "string"
            ]
          },
          "country": {
            "type": [
              "null",
              "string"
            ]
          },
          "state": {
            "type": [
              "null",
              "string"
            ]
          },
          "city": {
            "type": [
              "null",
              "string"
            ]
          },
          "annualRevenueUsd": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "createdDate": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "salesOrders": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SalesOrder"
            }
          }
        }
      },
      "OrderItem": {
        "type": "object",
        "properties": {
          "orderItemId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "orderId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "productId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "quantity": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "unitPriceUsd": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "lineTotalUsd": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "order": {
            "$ref": "#/components/schemas/SalesOrder"
          },
          "product": {
            "$ref": "#/components/schemas/Product"
          }
        }
      },
      "Product": {
        "type": "object",
        "properties": {
          "productId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "productName": {
            "type": "string"
          },
          "productCategory": {
            "type": [
              "null",
              "string"
            ]
          },
          "processNodeNm": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "packageType": {
            "type": [
              "null",
              "string"
            ]
          },
          "unitPriceUsd": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "lifecycleStatus": {
            "type": [
              "null",
              "string"
            ]
          },
          "createdDate": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "orderItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItem"
            }
          }
        }
      },
      "SalesOrder": {
        "type": "object",
        "properties": {
          "orderId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "customerId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "salesRepId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "orderDate": {
            "type": "string",
            "format": "date"
          },
          "orderStatus": {
            "type": [
              "null",
              "string"
            ]
          },
          "totalAmountUsd": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "currency": {
            "type": [
              "null",
              "string"
            ]
          },
          "customer": {
            "$ref": "#/components/schemas/Customer"
          },
          "orderItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItem"
            }
          },
          "salesRep": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/SalesRep"
              }
            ]
          }
        }
      },
      "SalesRep": {
        "type": "object",
        "properties": {
          "salesRepId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "repName": {
            "type": "string"
          },
          "region": {
            "type": [
              "null",
              "string"
            ]
          },
          "email": {
            "type": [
              "null",
              "string"
            ]
          },
          "hireDate": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "salesOrders": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SalesOrder"
            }
          }
        }
      },
      "VwSalesFact": {
        "type": "object",
        "properties": {
          "orderId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "orderDate": {
            "type": "string",
            "format": "date"
          },
          "customerName": {
            "type": "string"
          },
          "customerType": {
            "type": [
              "null",
              "string"
            ]
          },
          "productName": {
            "type": "string"
          },
          "productCategory": {
            "type": [
              "null",
              "string"
            ]
          },
          "processNodeNm": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "quantity": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "unitPriceUsd": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "lineTotalUsd": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "repName": {
            "type": [
              "null",
              "string"
            ]
          },
          "region": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "ProductDocument": {
        "type": "object",
        "properties": {
          "fileName": {
            "type": "string",
            "description": "The full blob name / path of the document"
          },
          "productName": {
            "type": "string",
            "description": "The product name this document belongs to"
          },
          "documentType": {
            "type": "string",
            "description": "The type of document: Engineering DataSheet, Marketing Brief, Sales One Pager, or Other"
          },
          "size": {
            "type": "integer",
            "format": "int64",
            "description": "File size in bytes"
          },
          "contentType": {
            "type": "string",
            "description": "The MIME content type of the document"
          },
          "lastModified": {
            "type": ["null", "string"],
            "format": "date-time",
            "description": "The last modified timestamp"
          },
          "url": {
            "type": "string",
            "description": "The direct URL to the blob"
          }
        }
      },
      "ProductDescription": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The product identifier (e.g. chip-1, Chip-100)"
          },
          "description": {
            "type": "string",
            "description": "The full product description text"
          }
        }
      }
    }
  },
...
}